bring the memoized subfield collector into the class
