+++ src/execution/execute.ts
@@ -191,12 +191,6 @@ export function executeSync(args: ExecutionArgs): ExecutionResult {
 }
 
 /**
- * Constructs a ExecutionContext object from the arguments passed to
- * execute, which we will pass throughout the other execution methods.
- *
- * Throws a GraphQLError if a valid execution context cannot be created.
- *
- * TODO: consider no longer exporting this function
  * @internal
  */
 export function validateExecutionArgs(
